+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFDXJJBWZVMBJJY9SETCVS2D"
+        "01KFDY65K88EQYPSFX567FECWQ"
     )
     r"""The unique identifier of the data source"""