+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFDXJJBRVHWXD559AV7EGNZJ"
+        "01KFDY65JY81SX4VJN0N6VZF1D"
     )
     r"""The unique identifier of the data source"""